roles/relay/files/requirements.txt
discord
pyyaml
rcon